[NFC] rename / move InstanceTunnel to CoreBundle (ApplicationController::instanceTunnelAction is CoreBundle, too)
